refactor(vt-daemon): group SSE files into transport/sse/
transport/ had 17 immediate children, exceeding the 15-child fanout limit.
Move the three server-only SSE files (eventSubscriptionHub, agentEventsSse,
terminalRegistrySse) into a transport/sse/ subdirectory. eventTypes.ts stays
in transport/ since it is imported directly by renderer-side webapp code.

Update all import paths in the 7 affected files.
