+++ packages/systems/vt-daemon/src/transport/wsSubscriberWiring.ts
@@ -20,7 +20,7 @@ import {
     type Subscriber,
     type SubscriberHandle,
     type TopicName,
-} from './eventSubscriptionHub.ts'
+} from './sse/eventSubscriptionHub.ts'
 
 interface WsSubscriberContext {
     readonly handle: SubscriberHandle
